pyproject.toml
[tool.poetry]
name = "recruitment-car"
version = "0.1.0"
description = ""
authors = ["fiedosiukr@gmail.com"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0"
python-decouple = "^3.5"
psycopg2-binary = "^2.9.2"
djangorestframework = "^3.13.1"
requests = "^2.27.1"
pydantic = "^1.9.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"

[tool.black]
line-length = 119

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
